Adding CSV endocing for headers and decode bug fix.
The JSON encoding was to verbose to store on GAE's datastore so a CSV
implementation keeps parsing relatively simple and reduces the encoded
size.

While working on this a problem decoding packets outside of the
library showed up.  Still not sure why pkt.Decode() was working on from
within the pcap pkg, but failing for long lists of packets when called
from outside the pcap pkg.  It may have something to do with the C
memory references.  Since decoding needed to be done internally we have
made it a private method and calling it when the packet is received.