Adding unit test for pkt/tcp.go.
Having the testHelp.go that contains the C data that we need to test our
cgo functions allows us to run unit tests.  This is the first of those
unit tests for TCP headers.

Also fixed a small bug in example/tcpdump.go where it was manually
calling the Decode function.