Adding UDP header support.
Implemented by Alessandro Guido <ag@alessandroguido.name>

This adds the ability to detect and decode UDP headers.  I also added a
verbose flag to the example tcpdump command to allow JSON encoded output
to we could do a quick test to ensure that UDP was being correctly
identified and decoded.