Adding access to C.pcap_set_buffer_size.
To use C.pcap_set_buffer_size one must first create a C.pcap_t so we
needed to add access to C.pcap_create among other functions used to set
the values that would have been set by the OpenLive call.  The reason to
undertake all this effort is that increasing the buffer size can greatly
reduce the number of packets for a trace.  Some functions were move in
the pcap.go source file with this change.

To illustrate these changes an example of creating a Pcap and setting
the options including the buffer size was added to the tcpdump example.