Removing dead code and hiding an unsafe pointer.
There was an unused field in the Packet struct that would have pointed
to possibly invalid C memory that has been removed, and the buf field
was not used outside of the pkt package and making it private is safer.

Also adding a quiet flag (-q) for the tcpdump example that allows no
per-packet output to the console and simply displays the stats at the end.