Adding trace package and refactoring code.
The Filter struct was not being used in the pcap package so it has been
moved to a sub-package of its own.  These filters can be added to to aid
in creating complex filter strings to pass as arguments to a Pcap, but
for now they are not that useful.

The Stat struct was moved to its own sub-package to clean up the pcap
package and so that they can be used without importing packages that
depend on cgo.

The trace package is being added as a way to archive and analyze pcap
data.  A PktTrace struct combines packet data with pcap meta data so
that it can be easily serialized as a gob and saved to a file or
transmitted to another server.  This package will also contain functions
for analyzing the trace data.

If you have code that uses pcap.Stat or pcap.Filter you should import
the new packages:
  code.google.com/p/golibpcap/pcap/stat
  code.google.com/p/golibpcap/pcap/filter
And substitute:
  pcap.Stat -> stat.Stat
  pcap.Filter -> filter.Filter

The new trace package is very much an experiment and may change without
warning.  These small refactoring changes should be minimally disruptive
and infrequent.