Factoring unsafe and cgo with build constraints.
With the goal of having trace.PktTrace be platform independent and
useful anywhere we can run Go code this change separates the source into
files that rely on pcap and cgo and those that do not.  To compile this
package to run on a system that does not have libpcap or cgo you should
use the "safe" tag as a build constraint
(http://golang.org/pkg/go/build/).  This change should also allow the
use of trace.PktTrace on appengine as it now supports the appengine build
constraint as well.

The example binary also supports the "safe" build tag and produces a
command that will read gzip compressed gob encoded trace.PktTrace
structs from a file.