Speeding up pkt.decode by reducing calls to append.
Since the most common case for packet capture is a packet that has link,
network, and transport OSI layers pre-allocating an []Hdr with three spots
and assigning to those known indices speeds up the decode process.
This also allows easier access to these common headers by applications
that are processing the data.  These indices are also included by name
to ensure accuracy.  This required a check for nil Hdr in Packet
serialization calls, but since this should ideally be done off-line it
seems like a good trade-off to make.

To aid the implementation of applications to work with the trace data we
have also added useful constants for decoding the encapsulated type of
the Ethernet Frame, IP protocol numbers and TCP flags.