Adding interface for IP headers.
A common issue when accessing the network level headers in a packet
trace is that IPv6 and IPv4 are distinct headers by commonly the
information needed from them is the same.  To solve this we have
introduced the InetProtoHdr interface that provides access to the source
IP, destination IP, protocol number and payload length.

This change allowed for some optimizations in the packet decoding
function that reduces the number of local variables used, GC cycles and
total malloc calls.  In evaluating these claims we also optimized the
example code by moving a declaration outside of a loop.