Fixing an offset error in pcap/pkt/tcp_unsafe.go.
If you are using TCP flags in your analysis you should apply this update
ASAP.  Unfortunately data that was collected prior to this fix will not
have the proper flag data in the pkt.TcpHdr.Flags field.

To get the TCP flags we should have indexed 12 octets into the packet
header instead of the 14 we were using.  This was causing the TCP flags
to be set by the window size.