Adding IP and TCP level trace analysis functions.
The new IPFlow and TCPFlow structs allow make the analysis of TCP/IP and
IP traffic easier.  We also have functions that given an []pkt.Packet
will return a map of address tuples to the packets that belong to that
tuple.  An example application would be picking a single client out of a
trace that contained many.

The safe tcpdump example has also been updated to include this
functionality.