Bug fix for packet time stamps.
The units are incorrect when converting the libpcap timeval to a Go
time.Time.  They were using microseconds and we were treating them as
nanoseconds.  The simple fix was to multiply by 1000.

To fix existing data the pkt.Packet.Time (ot) value should be converted
to the correct time (nt) as follows:

  s := ot.Unix() // recover number of seconds
  ms := ot.UnixNano() - (s * 1000000000) // recover ms given as ns
  nt := time.Unix(s, ms * 1000) // recreate time with ns proper

Any data that was created after this fix will not need this conversion.