Fixing a memory leak and API updates.
The pcap close() method was being used only for pcap file reading, but
it is actually used to free the buffer used for capturing packets as
well.  This change makes this method available to all pcap traces.

We also make some default values available in the pcap package.

In the experimental trace package we are renaming an awkwardly named
function and adding a constant for the archive file MIME type.