pcap savefiles OK to use and tracking libpcap version
When reading an pcap savefile we were missing being able to read the
snaplen, and saving the FileName to the trace.MetaPcap data.  We also
will not get stats from a savefile so we have to check for that in the
safe tcpdump example.

In the trace we have introduced the LibVersion and FileName so the
trace.Version string has been changed to beta-0-0-1.  We have also added
a few new error values that will be used in the new TCPFlow code.

Now that reading from savefiles is working any valid pcap savefile can
now be opened with and analyzed with the forthcoming analysis code.