Adding analysis code for TCP and HTTP packets.
The TCPFlow can filter out the packets from a trace that are relevant to
a particular TCP tuple and allows the analysis of those packets for the
number of reordered packets, losses and retransmissions.  Functions to
calculate the instantaneous throughput and capacity for the flow are in
this change, but need more work.

The HTTPExchange allows you to easily work with the packets that make
up a specific HTTP request and corresponding response.  The TCPFlow for
those packets can then be extracted from the exchange and analyzed.