Fixing a problem with the blocking pcap_breakloop.
pcap_breakloop will block while waiting on a packet to arrive after
it sets a flag to stop the loop.  This can be bad if you are calling
BreakLoop on a connection that has already disconnected.  For this
reason we are calling C.pcap_breakloop in another goroutine and
timing out after 1 ms.